% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta0.R
\name{beta0}
\alias{beta0}
\title{particular incomplete beta function}
\usage{
beta0(x0, x1)
}
\arguments{
\item{x0}{scalar with lower boundary for integration}

\item{x1}{scalar with upper boundary for integration}
}
\value{
scalar with particular incomple beta function
}
\description{
\deqn{B_x1(4/3,0) - B_x0(4/3,0) = \int_x0^x1 t^(4/3-1) (1-t)^(-1) dt}{B_x1(4/3,0) - B_x0(4/3,0) = int_x0^x1 t^(4/3-1) (1-t)^(-1) dt}
}
\examples{
beta0(0.1, 0.2)

}
