% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dget_l_ISO.R
\name{dget_l_ISO}
\alias{dget_l_ISO}
\title{Differential equation}
\usage{
dget_l_ISO(vH, l, pars)
}
\arguments{
\item{vH}{scalar with scaled maturity}

\item{l}{scalar with scaled length}

\item{pars}{5-vector k, lT g f sM}
}
\value{
dl: scalar with d/dvH l
}
\description{
Differential equation
}
