% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lj.R
\name{get_lj}
\alias{get_lj}
\title{Gets scaled length at metamorphosis}
\usage{
get_lj(p, f, lb0)
}
\arguments{
\item{p}{6-vector with parameters: g, k, l_T, v_H^b, v_H^j, v_H^p; if p is a 5-vector, output lp is empty}

\item{f}{optional scalar with scaled functional responses (default 1)}

\item{lb0}{optional scalar with scaled length at birth}
}
\value{
[lj, lp, lb, info]  lj: scalar with scaled length at metamorphosis; * lp: scalar with scaled length at puberty; lb: scalar with scaled length at birth;  info: indicator equals 1 if successful
}
\description{
Type M-acceleration: Isomorph, but V1-morph between vHb and vHj
This routine obtaines scaled length at metamorphosis lj given scaled muturity at metamorphosis vHj.
The theory behind get_lj, is discussed in the comments to DEB3.
If scaled length at birth (third input) is not specified, it is computed (using automatic initial estimate);
 if it is specified. however, is it just copied to the (third) output.
The code assumes vHb < vHj < vHp (see first input).
}
