% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lp.R
\name{get_lp}
\alias{get_lp}
\title{Get scaled length at puberty}
\usage{
get_lp(p, f = 1, lb0 = NULL)
}
\arguments{
\item{p}{5-vector with parameters: g, k, l_T, v_H^b, v_H^p}

\item{f}{optional scalar with scaled functional responses (default 1)}

\item{lb0}{optional scalar with scaled length at birth or optional 2-vector with scaled length, l,
and scaled maturity, vH, for a juvenile that is now exposed to f, but previously at another f. lb0 should be specified for foetal development.}
}
\value{
3-vector: lp: scalar with scaled length at puberty; lb: scalar with scaled length at birth;  info: indicator equals 1 if successful, 0 otherwise
}
\description{
Get scaled length at puberty
}
\examples{
get_lp(c(.5, .1, .1, .01, .2))
}
