% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tb.R
\name{get_tb}
\alias{get_tb}
\title{get_tb}
\usage{
get_tb(p, eb = 1, lb = NA)
}
\arguments{
\item{p}{1 or 3-vector with parameters g, k_J/ k_M, v_H^b; Last 2 values are optional in invoke call to get_lb}

\item{eb}{optional scalar with scaled reserve density at birth (default eb = 1)}

\item{lb}{optional scalar with scaled length at birth (default: lb is obtained from get_lb)}
}
\value{
3-element vector containing tb: scaled age at birth tau_b = a_b k_M; lb: scalar with scaled length at birth: L_b/ L_m; info: indicator equals 1 if successful, 0 otherwise
}
\description{
Obtains scaled age at birth, given the scaled reserve density at birth.
Divide the result by the somatic maintenance rate coefficient to arrive at age at birth.
}
\examples{
get_tb(p = c(.1,.5,.03))
}
\author{
Bas Kooijman
}
