% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tm_s.R
\name{get_tm_s}
\alias{get_tm_s}
\title{Obtains scaled mean age at death fro short growth periods}
\usage{
get_tm_s(p, f, lb, lp)
}
\arguments{
\item{p}{4 or 7-vector with parameters: [g lT ha sG] or [g k lT vHb vHp ha SG]}

\item{f}{optional scalar with scaled reserve density at birth (default eb = 1)}

\item{lb}{optional scalar with scaled length at birth (default: lb is obtained from get_lb)}

\item{lp}{optional scalar with scaled length at puberty}
}
\value{
[tm, Sb, Sp, info]
* tm: scalar with scaled mean life span
* Sb: scalar with survival probability at birth (if length p = 7)
* Sp: scalar with survival prabability at puberty (if length p = 7)
* info: indicator equals 1 if successful, 0 otherwise
}
\description{
Obtains scaled mean age at death assuming a short growth period relative to the life span
Divide the result by the somatic maintenance rate coefficient to arrive at the mean age at death.
The variant get_tm_foetus does the same in case of foetal development.
If the input parameter vector has only 4 elements (for [g, lT, ha/ kM2, sG]),
  it skips the calulation of the survival probability at birth and puberty.
}
