% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tp.R
\name{get_tp}
\alias{get_tp}
\title{Gets scaled age at puberty}
\usage{
get_tp(p, f = 1, lb0 = NULL)
}
\arguments{
\item{p}{5-vector with parameters: g, k, l_T, v_H^b, v_H^p}

\item{f}{optional scalar with functional response (default f = 1)}

\item{lb0}{optional scalar with scaled length at birth or optional 2-vector with scaled length, l, and scaled maturity, vH, for a juvenile that is now exposed to f, but previously at another f}
}
\value{
[tp, tb, lp, lb, info]
}
\description{
Gets scaled age at puberty
}
\examples{
get_tp(c(.5, .1, .1, .01, .2))
}
