% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_v_Hb.R
\name{get_v_Hb}
\alias{get_v_Hb}
\title{Scaled maturity volume at birth}
\usage{
get_v_Hb(E_Hb, kap, g, p_Am, v, L_m)
}
\arguments{
\item{E_Hb}{maturity at birth}

\item{kap}{kappa}

\item{g}{energy investment ratio}

\item{p_Am}{maximum specific assimilation flux}

\item{v}{energy conductance}

\item{L_m}{maximum structural length}
}
\value{
Scaled maturity volume at birth v_Hb scalar
}
\description{
Scaled maturity volume at birth
}
\author{
Philipp Boersch-Supan
}
