% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmregr.R
\name{nmregr}
\alias{nmregr}
\title{Calculates least squares estimates using Nelder Mead's simplex method}
\usage{
nmregr(func, p, varargin)
}
\arguments{
\item{func}{string with name of user-defined function}

\item{p}{(k,2)-matrix with p(:,1) initial guesses for parameter values;  p(:,2) binaries with yes or no iteration (optional)}

\item{varargin}{MATLAB idiom...}
}
\value{
[q, info]
}
\description{
Calculates least squares estimates using Nelder Mead's simplex method
}
